Create MW4Runner-ver2_4
This is the binary file for the Runner 250 R for the MultiWii version 2.4
This is based on Tom Hogue work.
This binary can be use directly with the Walkera "FlyConfig.exe" application.
This binary will work with the Walkera Boot Loader. It will allow to go back to version 2.2 if you want to.